% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec_scales.R
\name{dec_pal}
\alias{dec_pal}
\title{december palettes with ramped colours
This code is heavily based on the set up of the ochRe package: https://github.com/ropenscilabs/ochRe/blob/master/R/colors.R}
\usage{
dec_pal(palette = "winter_solstice", alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Choose from 'december_palettes' list}

\item{alpha}{transparency}

\item{reverse}{If TRUE, the direction of the colours is reversed.}
}
\description{
december palettes with ramped colours
This code is heavily based on the set up of the ochRe package: https://github.com/ropenscilabs/ochRe/blob/master/R/colors.R
}
\examples{
library(scales)
show_col(dec_pal()(10))

filled.contour(volcano,color.palette = dec_pal(), asp=1)

}
