% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec_colours.R
\docType{data}
\name{december_palettes}
\alias{december_palettes}
\title{Colour Palettes for December Holidays
This code is heavily based on the set up of the ochRe package: https://github.com/ropenscilabs/ochRe/blob/master/R/colors.R}
\format{
An object of class \code{list} of length 6.
}
\usage{
december_palettes
}
\description{
A collection of colour palettes for a range of December holidays. The list of available palettes is:
hanukkah_rainbow
hanukkah_simple
xmas
yulelog
kwanzaa
winter_solstice
}
\examples{

# Make an x-y plot using the winter_solstice palette
library(tidyverse)
df <- data.frame(x = rnorm(100, 0, 20),
          y = rnorm(100, 0, 20),
          cl = sample(letters[1:8], 100, replace=TRUE))
ggplot(df, aes(x, y, colour=cl, shape=cl)) +
  geom_point(size=4) + scale_colour_december() +
  theme_bw() + theme(aspect.ratio=1)

# Make a histogram using the McCrea Collins Street palette
ggplot(df, aes(x, fill=cl)) + geom_histogram() +
  scale_fill_december(palette="hanukkah_rainbow")

}
\keyword{datasets}
